Cleanup event list (#45)
