+++ web/src/pages/EventListPage/EventListPage.scss
@@ -1,9 +1,10 @@
-.EventListPage {
-  h2 {
+.eventListPage {
+  h1 {
     margin-left: 10px;
+    color: #000;
   }
 
   aside {
-    margin-top: 60px;
+    margin-top: 73px;
   }
 }
