+++ web/src/pages/EventListPage/EventListPage.tsx
@@ -4,20 +4,37 @@ import EventList from "../../components/event/EventList";
 import Header from "../../components/common/Header";
 import Footer from "../../components/common/Footer";
 import EventCalendar from "../../components/event/EventCalendar";
+import { withRouter, RouteComponentProps } from "react-router-dom";
 import "./EventListPage.scss";
 
-class EventListPage extends Component {
+class EventListPage extends Component<RouteComponentProps<any>> {
   state = {
-    pastEvents: [],
-    futureEvents: []
+    type: "upcoming",
+    events: []
   };
 
   async componentDidMount() {
-    const eventService = new EventService();
-    const pastEvents = await eventService.getEvents();
-    const futureEvents = await eventService.getEvents();
+    this.fetch();
+  }
 
-    this.setState({ pastEvents, futureEvents });
+  async componentDidUpdate() {
+    if (
+      this.props.match.params.type &&
+      this.state.type !== this.props.match.params.type
+    ) {
+      this.fetch();
+    }
+  }
+
+  async fetch() {
+    const eventService = new EventService();
+    const type = this.props.match.params.type || "upcoming";
+    const events = await eventService.getEvents(type);
+    this.setState({ events, type });
+  }
+
+  capitalizeFirstLetter(word: string) {
+    return word.charAt(0).toUpperCase() + word.slice(1);
   }
 
   render() {
@@ -29,10 +46,14 @@ class EventListPage extends Component {
 
         <div className="flexContainer wrapper">
           <main>
-            <h2>Next Events</h2>
-            <EventList events={this.state.futureEvents} />
-            <h2>Past Events</h2>
-            <EventList events={this.state.pastEvents} />
+            {this.state.events.length > 0 ? (
+              <div>
+                <h1>{this.capitalizeFirstLetter(this.state.type)} Events</h1>
+                <EventList events={this.state.events} />
+              </div>
+            ) : (
+              ""
+            )}
           </main>
 
           <aside>
@@ -46,4 +67,4 @@ class EventListPage extends Component {
   }
 }
 
-export default EventListPage;
+export default withRouter(EventListPage);
